requirements.txt
cronex >= 0.1.3.1
requests >= 2.31.0
aprslib >= 0.7.2
pySerial >= 3.4